+++ Model/CurrentReply.cs
@@ -8,10 +8,10 @@ namespace Model
 {
     public class CurrentReply
     {
-        public int      Id          { set; get; }
-        public int      CurrentId   { set; get; }
-        public int      AimsId      { set; get; }
-        public int      UserId      { set; get; }
+        public long     Id          { set; get; }
+        public long     CurrentId   { set; get; }
+        public long     AimsId      { set; get; }
+        public long     UserId      { set; get; }
         public string   Contents    { set; get; }
         public DateTime ReplyDate   { set; get; }
         public int      PraiseCount { set; get; }
