+++ Model/NoticeTrace.cs
@@ -8,9 +8,9 @@ namespace Model
 {
     public class NoticeTrace
     {
-        public int      Id        { set; get; }
-        public int      NoticeId  { set; get; }
-        public int      UserId    { set; get; }
+        public long     Id        { set; get; }
+        public long     NoticeId  { set; get; }
+        public long     UserId    { set; get; }
         public DateTime TraceDate { set; get; }
     }
 }
